requirements.txt
streamlit 
google-auth 
google-auth-oauthlib 
google-api-python-client 
langchain 
langchain_community
langchain_huggingface
transformers
torch
langgraph 
openai 
huggingface_hub